requirements.txt
Django==1.11
django-filter==1.1.0
django-rest-framework==0.1.0
djangorestframework==3.7.7
psycopg2==2.7.3.2
pytz==2017.3
